Renomeando nome do projeto e pacote entities
