--- src/main/java/ProjetoMensalRestaurante/Entity/Cardapio.java
@@ -1,4 +1,0 @@
-package ProjetoMensalRestaurante.Entity;
-
-public class Cardapio {
-}
