--- src/main/java/ProjetoMensalRestaurante/Entity/Cliente.java
@@ -1,4 +1,0 @@
-package ProjetoMensalRestaurante.Entity;
-
-public class Cliente {
-}
