--- src/main/java/ProjetoMensalRestaurante/Entity/Funcionario.java
@@ -1,4 +1,0 @@
-package ProjetoMensalRestaurante.Entity;
-
-public class Funcionario {
-}
