--- src/main/java/ProjetoMensalRestaurante/Entity/Ingredientes.java
@@ -1,4 +1,0 @@
-package ProjetoMensalRestaurante.Entity;
-
-public class Ingredientes {
-}
