--- src/main/java/ProjetoMensalRestaurante/Entity/Pedido.java
@@ -1,4 +1,0 @@
-package ProjetoMensalRestaurante.Entity;
-
-public class Pedido {
-}
