+++ src/main/java/com/dove/Main.java
@@ -1,4 +1,4 @@
-package ProjetoMensalRestaurante;
+package com.dove;
 
 public class Main {
     public static void main(String[] args) {
