+++ src/main/java/com/dove/entities/Cardapio.java
@@ -1,0 +1,4 @@
+package com.dove.entities;
+
+public class Cardapio {
+}
