+++ src/main/java/com/dove/entities/Ingredientes.java
@@ -1,0 +1,4 @@
+package com.dove.entities;
+
+public class Ingredientes {
+}
