+++ src/main/java/com/dove/entities/Pedido.java
@@ -1,0 +1,4 @@
+package com.dove.entities;
+
+public class Pedido {
+}
